refactor(v1): drop oai_ prefix from the chat_completions dialect module
