+++ verifiers/v1/dialects/__init__.py
@@ -1,13 +1,13 @@
 """Wire dialects: per-native-format translators (wire -> vf) for the interception trace.
 
-`Dialect` is the abstraction; `oai_chat_completions` is the only dialect today (OpenAI Responses
+`Dialect` is the abstraction; `chat_completions` is the only dialect today (OpenAI Responses
 / Anthropic Messages become new modules). The interception server serves each registered
 dialect's `routes`, so the wire format is resolved from the endpoint the program's SDK posts to
 — no per-harness declaration.
 """
 
 from verifiers.v1.dialects.base import Dialect
-from verifiers.v1.dialects.oai_chat_completions import (
+from verifiers.v1.dialects.chat_completions import (
     FINISH_REASONS,
     ChatCompletionsDialect,
     parse_message,
